+++ sprockets/mixins/http/__init__.py
@@ -9,7 +9,6 @@ import asyncio
 import logging
 import os
 import time
-from urllib import parse
 
 from ietfparse import algorithms, errors, headers
 from sprockets.mixins.mediatype import transcoders
@@ -30,7 +29,6 @@ class HTTPResponse:
      :meth:`~sprockets.mixins.http.HTTPClientMixin.http_fetch` method.
 
     """
-
     def __init__(self, simplify_error_response=True):
         self._exceptions = []
         self._finish = None
@@ -64,11 +62,10 @@ class HTTPResponse:
         """
         self._responses.append(response)
         if 'Warning' in response.headers:
-            LOGGER.warning(
-                'HTTP %s %s Warning (%s): %s (attempt %s)',
-                response.request.method, response.request.url,
-                response.code, response.headers['Warning'],
-                len(self._responses))
+            LOGGER.warning('HTTP %s %s Warning (%s): %s (attempt %s)',
+                           response.request.method, response.request.url,
+                           response.code, response.headers['Warning'],
+                           len(self._responses))
 
     def finish(self):
         """Mark the processing as finished"""
@@ -197,8 +194,7 @@ class HTTPResponse:
         if isinstance(value, list):
             return [self._decode(v) for v in value]
         elif isinstance(value, dict):
-            return {self._decode(k): self._decode(v)
-                    for k, v in value.items()}
+            return {self._decode(k): self._decode(v) for k, v in value.items()}
         elif isinstance(value, bytes):
             return value.decode('utf-8')
         return value
@@ -215,10 +211,10 @@ class HTTPResponse:
         if 'Content-Type' not in self._responses[-1].headers:
             return self._responses[-1].body
         try:
-            content_type = algorithms.select_content_type(
-                [headers.parse_content_type(
-                    self._responses[-1].headers['Content-Type'])],
-                AVAILABLE_CONTENT_TYPES)
+            content_type = algorithms.select_content_type([
+                headers.parse_content_type(
+                    self._responses[-1].headers['Content-Type'])
+            ], AVAILABLE_CONTENT_TYPES)
         except errors.NoMatch:
             return self._responses[-1].body
 
@@ -226,8 +222,8 @@ class HTTPResponse:
             return self._decode(
                 self._json.loads(self._decode(self._responses[-1].body)))
         elif content_type[0] == CONTENT_TYPE_MSGPACK:  # pragma: nocover
-            return self._decode(
-                self._msgpack.unpackb(self._responses[-1].body))
+            return self._decode(self._msgpack.unpackb(
+                self._responses[-1].body))
 
     def _error_message(self):
         """Try and extract the error message from a HTTP error response.
@@ -258,7 +254,8 @@ class HTTPClientMixin:
         self.__hcm_msgpack = transcoders.MsgPackTranscoder()
         self.simplify_error_response = True
 
-    async def http_fetch(self, url,
+    async def http_fetch(self,
+                         url,
                          method='GET',
                          request_headers=None,
                          body=None,
@@ -267,6 +264,7 @@ class HTTPClientMixin:
                          max_redirects=None,
                          connect_timeout=None,
                          request_timeout=None,
+                         retry_timeout=None,
                          max_http_attempts=None,
                          auth_username=None,
                          auth_password=None,
@@ -277,7 +275,10 @@ class HTTPClientMixin:
                          **kwargs):
         """Perform a HTTP request
 
-        Will retry up to ``self.MAX_HTTP_RETRIES`` times.
+        Will retry up to `max_http_attempts` times with an exponentially
+        increasing sleep time starting with `retry_timeout` seconds.  If
+        a ``Retry-Header`` is included in a response, then it will override
+        the calculated sleep time.
 
         :param str url: The URL for the request
         :param str method: The HTTP request method, defaults to ``GET``
@@ -294,6 +295,8 @@ class HTTPClientMixin:
             seconds, default 20 seconds
         :param float request_timeout:  Timeout for entire request in seconds,
             default 20 seconds
+        :param float retry_timeout:  Time to sleep between retries,
+            default 3 seconds
         :param int max_http_attempts: Maximum number of times to retry
             a request, default is 3 attempts
         :param str auth_username: Username for HTTP authentication
@@ -314,6 +317,7 @@ class HTTPClientMixin:
             is specified
 
         """
+
         # Curry the request parameters through from our named params
         def apply_default(val, default):
             return default if val is None else val
@@ -323,6 +327,8 @@ class HTTPClientMixin:
                                         self.DEFAULT_CONNECT_TIMEOUT)
         request_timeout = apply_default(request_timeout,
                                         self.DEFAULT_REQUEST_TIMEOUT)
+        retry_timeout = apply_default(retry_timeout,
+                                      self.DEFAULT_RETRY_TIMEOUT)
         max_http_attempts = apply_default(max_http_attempts,
                                           self.MAX_HTTP_RETRIES)
 
@@ -365,18 +371,16 @@ class HTTPClientMixin:
                                'called with raise_error')
 
         for attempt in range(0, max_http_attempts):
-            LOGGER.debug('%s %s (Attempt %i of %i) %r',
-                         method, url, attempt + 1, max_http_attempts,
-                         request_headers)
+            LOGGER.debug('%s %s (Attempt %i of %i) %r', method, url,
+                         attempt + 1, max_http_attempts, request_headers)
             if attempt > 0:
                 request_headers['X-Retry-Attempt'] = str(attempt + 1)
             try:
-                resp = await client.fetch(
-                    str(url),
-                    headers=request_headers,
-                    body=body,
-                    raise_error=False,
-                    **kwargs)
+                resp = await client.fetch(str(url),
+                                          headers=request_headers,
+                                          body=body,
+                                          raise_error=False,
+                                          **kwargs)
             except (OSError, httpclient.HTTPError) as error:
                 response.append_exception(error)
                 LOGGER.warning(
@@ -393,29 +397,35 @@ class HTTPClientMixin:
                 return response
             elif resp.code in dont_retry:
                 break
-            elif resp.code in {423, 429, 503}:
-                await self._http_resp_rate_limited(
-                    resp, min(connect_timeout, request_timeout))
-            elif resp.code < 500:
-                LOGGER.debug('HTTP Response Error for %s to %s'
-                             'attempt %i of %i (%s): %s',
-                             method, url, resp.code, attempt + 1,
-                             max_http_attempts, response.body)
+            elif resp.code < 500 and resp.code not in {423, 429}:
+                LOGGER.debug(
+                    'HTTP Response Error for %s to %s'
+                    'attempt %i of %i (%s): %s', method, url, resp.code,
+                    attempt + 1, max_http_attempts, response.body)
                 response.finish()
                 return response
 
             LOGGER.warning(
-                'HTTP Error for %s to %s, attempt %i of %i (%s): %s',
-                method, url, attempt + 1, max_http_attempts, resp.code,
-                response.body)
+                'HTTP Error for %s to %s, attempt %i of %i (%s): %s', method,
+                url, attempt + 1, max_http_attempts, resp.code, response.body)
+
+            if attempt + 1 != max_http_attempts:
+                if response.headers.get('Retry-After'):
+                    retry_after = min(int(response.headers['Retry-After']),
+                                      request_timeout)
+                else:
+                    retry_after = (2**attempt) * retry_timeout
+                LOGGER.debug('Sleeping for %.f seconds before retry',
+                             retry_after)
+                await asyncio.sleep(retry_after)
 
         LOGGER.warning('HTTP %s to %s failed after %i attempts', method, url,
                        max_http_attempts)
         response.finish()
         return response
 
-    def _http_req_apply_default_headers(self, request_headers,
-                                        content_type, body):
+    def _http_req_apply_default_headers(self, request_headers, content_type,
+                                        body):
         """Set default values for common HTTP request headers
 
         :param dict request_headers: The HTTP request headers
@@ -432,14 +442,14 @@ class HTTPClientMixin:
             'Accept', ', '.join([str(ct) for ct in AVAILABLE_CONTENT_TYPES]))
         if body:
             request_headers.setdefault(
-                'Content-Type', str(content_type) or str(CONTENT_TYPE_MSGPACK))
+                'Content-Type',
+                str(content_type) or str(CONTENT_TYPE_MSGPACK))
         if hasattr(self, 'correlation_id'):
-            request_headers.setdefault(
-                'Correlation-Id', self.correlation_id)
+            request_headers.setdefault('Correlation-Id', self.correlation_id)
         elif hasattr(self, 'request') and \
                 self.request.headers.get('Correlation-Id'):
-            request_headers.setdefault(
-                'Correlation-Id', self.request.headers['Correlation-Id'])
+            request_headers.setdefault('Correlation-Id',
+                                       self.request.headers['Correlation-Id'])
         return request_headers
 
     def _http_req_body_serialize(self, body, content_type):
@@ -472,34 +482,16 @@ class HTTPClientMixin:
         """
         # Tornado Request Handler
         try:
-            return '{}/{}'.format(
-                self.settings['service'], self.settings['version'])
+            return '{}/{}'.format(self.settings['service'],
+                                  self.settings['version'])
         except (AttributeError, KeyError):
             pass
 
         # Rejected Consumer
         if hasattr(self, '_process'):
             try:
-                return '{}/{}'.format(
-                    self._process.consumer_name,
-                    self._process.consumer_version)
+                return '{}/{}'.format(self._process.consumer_name,
+                                      self._process.consumer_version)
             except AttributeError:
                 pass
         return DEFAULT_USER_AGENT
-
-    def _http_resp_rate_limited(self, response, timeout):
-        """Extract the ``Retry-After`` header value if the request was rate
-        limited and return a future to sleep for the specified duration.
-
-        :param tornado.httpclient.HTTPResponse response: The response
-        :param float timeout: Maximum number of seconds to wait regardless
-            of ``Retry-After`` header
-        :rtype: tornado.concurrent.Future
-
-        """
-        parsed = parse.urlparse(response.request.url)
-        duration = int(
-            response.headers.get('Retry-After', self.DEFAULT_RETRY_TIMEOUT))
-        LOGGER.warning('Rate Limited by %s, retrying in %i seconds',
-                       parsed.netloc, duration)
-        return asyncio.sleep(min(duration, timeout))